added Websocket, Message parsing, host change tracking, UI cleanup and source maps.
Websocket
===================
Connects to localhost on `DOMContentLoaded` sends a `request` command
and passes incoming messages to be parsed.

Parsing
==================
Creates an object with the ip as a key and the corresponding scan results
as the value.

Host change tracking
=================
Results are saved in local storage and then compered, hosts are added or updated
as needed.